backend/requirements.txt
langchain
langchain-openai
langchain-community
langchain-core
chromadb
tiktoken
openai
python-dotenv
pypdf
python-multipart
uvicorn
fastapi
